Validaciones de menú configuraciones, confirmación eliminar y regresar al inicio cuando se elimine la cuenta

+++ Moviles/app/src/main/java/mx/edu/plannert/MenuPerfilCuenta.kt
@@ -1,5 +1,6 @@
 package mx.edu.plannert
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -7,7 +8,9 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.database.*
 
 // TODO: Rename parameter arguments, choose names that match
@@ -21,7 +24,6 @@ private const val ARG_PARAM2 = "param2"
  * create an instance of this fragment.
  */
 class MenuPerfilCuenta : Fragment() {
-    // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
@@ -37,7 +39,6 @@ class MenuPerfilCuenta : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_menu_perfil_cuenta, container, false)
     }
 
@@ -52,40 +53,72 @@ class MenuPerfilCuenta : Fragment() {
 
     private fun eliminarCuenta() {
         val auth: FirebaseAuth = FirebaseAuth.getInstance()
-        val currentUser = auth.currentUser
+        val currentUser: FirebaseUser? = auth.currentUser
 
         if (currentUser != null) {
             val email = currentUser.email
 
-            // Eliminar usuario de la autenticación
-            currentUser.delete()
-                .addOnCompleteListener { authTask ->
-                    if (authTask.isSuccessful) {
-                        // Eliminar registro del usuario en Firebase Realtime Database
-                        val database: FirebaseDatabase = FirebaseDatabase.getInstance()
-                        val usuariosRef: DatabaseReference = database.getReference("usuarios")
+            val confirmationDialog = AlertDialog.Builder(requireContext())
+                .setTitle("Confirmación")
+                .setMessage("¿Estás seguro de que deseas eliminar tu cuenta?")
+                .setPositiveButton("Eliminar") { dialog, _ ->
+                    dialog.dismiss()
 
-                        usuariosRef.orderByChild("email").equalTo(email).addListenerForSingleValueEvent(object :
-                            ValueEventListener {
-                            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                                for (usuarioSnapshot in dataSnapshot.children) {
-                                    usuarioSnapshot.ref.removeValue()
-                                }
+                    // Eliminar usuario de la autenticación
+                    currentUser.delete()
+                        .addOnCompleteListener { authTask ->
+                            if (authTask.isSuccessful) {
+                                // Cerrar sesión y regresar al MainActivity
+                                auth.signOut()
+                                val intent = Intent(requireContext(), MainActivity::class.java)
+                                intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
+                                startActivity(intent)
+                                requireActivity().finish()
 
-                                // El usuario y su registro se eliminaron correctamente
-                                Toast.makeText(requireContext(), "¡Cuenta eliminada exitosamente!", Toast.LENGTH_SHORT).show()
+                                // Eliminar registro del usuario en Firebase Realtime Database
+                                val database: FirebaseDatabase = FirebaseDatabase.getInstance()
+                                val usuariosRef: DatabaseReference = database.getReference("usuarios")
+
+                                usuariosRef.orderByChild("email").equalTo(email)
+                                    .addListenerForSingleValueEvent(object : ValueEventListener {
+                                        override fun onDataChange(dataSnapshot: DataSnapshot) {
+                                            for (usuarioSnapshot in dataSnapshot.children) {
+                                                usuarioSnapshot.ref.removeValue()
+                                            }
+
+                                            // El usuario y su registro se eliminaron correctamente
+                                            Toast.makeText(
+                                                requireContext(),
+                                                "¡Cuenta eliminada exitosamente!",
+                                                Toast.LENGTH_SHORT
+                                            ).show()
+                                        }
+
+                                        override fun onCancelled(databaseError: DatabaseError) {
+                                            // Manejar el caso de error en la consulta de la base de datos
+                                            Toast.makeText(
+                                                requireContext(),
+                                                "¡Ha ocurrido un error al eliminar la cuenta!",
+                                                Toast.LENGTH_SHORT
+                                            ).show()
+                                        }
+                                    })
+                            } else {
+                                // Manejar el caso de error al eliminar la cuenta en la autenticación
+                                Toast.makeText(
+                                    requireContext(),
+                                    "¡Ha ocurrido un error al eliminar la cuenta!",
+                                    Toast.LENGTH_SHORT
+                                ).show()
                             }
-
-                            override fun onCancelled(databaseError: DatabaseError) {
-                                // Manejar el caso de error en la consulta de la base de datos
-                                Toast.makeText(requireContext(), "¡Ha ocurrido un error al eliminar la cuenta!", Toast.LENGTH_SHORT).show()
-                            }
-                        })
-                    } else {
-                        // Manejar el caso de error al eliminar la cuenta en la autenticación
-                        Toast.makeText(requireContext(), "¡Ha ocurrido un error al eliminar la cuenta!", Toast.LENGTH_SHORT).show()
-                    }
+                        }
                 }
+                .setNegativeButton("Cancelar") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .create()
+
+            confirmationDialog.show()
         } else {
             // Manejar el caso de usuario no autenticado
             Toast.makeText(requireContext(), "¡No se ha iniciado sesión!", Toast.LENGTH_SHORT).show()
